+++ src/renderer/styles/index.less
@@ -63,7 +63,7 @@ img {
 	font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
 	font-size: 14px;
 	-webkit-app-region: drag;
-	@apply h-8 flex w-full justify-between items-center relative select-none text-gray-800 dark:text-gray-200 bg-white dark:bg-gray-800;
+	@apply h-8 flex w-full justify-between items-center relative select-none text-gray-800 dark:text-gray-200;
 	.titlebar--icon {
 		@apply h-5;
 	}
