+++ src/renderer/vite-env.d.ts
@@ -2,25 +2,7 @@
 
 declare const electron: ElectronAPI;
 declare const APIROOT: string;
-declare const DEVELOPMENT: boolean;
 declare const VERSION: string;
-
-declare type ErrorMessage = `${ number } ${ string }`;
-
-declare interface APIErrorResponse {
-	success: false;
-
-	/* The error code */
-	message: ErrorMessage;
-
-	/* What went wrong on an API level */
-	description: string;
-
-	/* Human readable version of what happened (safe to show the user) */
-	readable?: string;
-}
-
-declare type APIResponse<T = unknown> = T & { success: true } | APIErrorResponse;
 
 declare const config: {
 	get(key: string): unknown;
